Adapt tests to EF Core 11, and drop two unrunnable green-test entries
Working through the tests that passed under EF 10 and no longer do. The green
list is an EF 10 artefact — no EF 11 run has been clean enough to update it — so
these are the EF 10 -> EF 11 delta rather than recent regressions.

Simple_decimal_literals_are_parsed_for_HasDefaultValue built its expectation
with (decimal)-1.1111, a runtime double->decimal cast, so preview 7 moved the
EXPECTATION to -1.1110999999999999765520897199 while scaffolding kept parsing
the literal exactly. Now a decimal literal. Upstream's SQL Server test has the
same latent bug and will hit it when they take the preview 7 SDK; they are
pinned to preview 6.

JsonTypes collection tests passed SQL Server's store types (nchar(32),
varchar(max)); Jet text is always Unicode and 255 is the varchar ceiling, so the
element facets asserted a type the mapping had already resolved. LibRed already
used the Jet values.

CompiledModel needed ManyTypes.Decimal's DecimalTypeDefaultWarning suppressed,
exactly as the SQL Server test does — same validator, same warning.

Select_DTO_constructor_distinct_..._after_client_eval no longer throws on Jet, so
the Assert.ThrowsAsync<TrueException> wrapper (which upstream keeps, because it
still fails on SQL Server) had to go, and its ORDER BY gained the collection key.
ToQueryString_for_include_reference_and_collection lost a now-redundant ordering
term.

Column_collection_of_nullable_strings_contains_null was overwritten during the
preview 7 sync with SQL Server's OPENJSON body. Jet has never been able to query
into a JSON collection: the EF 10 override asserted the translation failure, and
that is restored — calling its own base rather than the non-nullable one, which
the EF 10 version got wrong.

Also removed StringTranslationsJetTest.Equals and six ModelBuilding entries from
the green lists. Those tests are discovered but never execute — name collisions
with object.Equals and with protected generic helpers — so they can never pass
again and would gate the list forever. Fixed upstream by dotnet/efcore#38766 for
rc1, at which point the ModelBuilding ones return under the same names while
Equals becomes Instance_Equals.

+++ test/EFCore.Jet.FunctionalTests/Scaffolding/JetDatabaseModelFactoryTest.cs
@@ -1163,7 +1163,7 @@ namespace EntityFrameworkCore.Jet.FunctionalTests.Scaffolding
 
                     var column = columns.Single(c => c.Name == "A");
                     Assert.Equal("-1.1111", column.DefaultValueSql);
-                    Assert.Equal((decimal)-1.1111, column.DefaultValue);
+                    Assert.Equal(-1.1111m, column.DefaultValue);
 
                     column = columns.Single(c => c.Name == "B");
                     Assert.Equal("0.0", column.DefaultValueSql);
